+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project are documented in this file.
 
 ## 1.4.1 - Bug Fixes
 
--  Fixed Z-index issue causing inccorect element layering. 
+- Fixed Z-index issue causing incorrect element layering (pill buttons now render behind
+- Added Extension Settings option to position ModBox Queues Box at bottom-left or bottom-right 
 
 ## 1.4.0 - ModBox Links